gateway添加多设备校验，一些bug修复

+++ server/nov-starter/nov-service/service_edu/src/main/java/com/novedu/nov/edu/entity/vo/DashBoardInfoVO.java
@@ -3,6 +3,7 @@ package com.novedu.nov.edu.entity.vo;
 import com.novedu.nov.edu.entity.EduCourse;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -17,7 +18,7 @@ public class DashBoardInfoVO {
 
     private Integer OrderCount;
 
-    private double orderAmount;
+    private BigDecimal orderAmount;
 
     private String teacherName;
 }
